refactor: extract ChatService domain service and centralize prompts
- Centralize all GPT prompts and context templates into `App\Domain\Bot\Messages`.
- Extract core AI response logic from `DefaultChatHandler` into a new `App\Domain\Bot\Service\ChatService`.
- Simplify `DefaultChatHandler` to delegate answer generation to `ChatService`.
- Modernize `CommandAndTriggerService::judgeCommand` using `Command::tryFrom()`.
- Update `Container` and `CommandHandlerFactory` for the new service wiring.
- Update existing tests and add `ChatServiceTest` to ensure 100% pass rate.
