+++ src/Application/CommandHandler/DefaultChatHandler.php
@@ -8,32 +8,21 @@ use App\Domain\Bot\ValueObject\Command;
 use App\Domain\Bot\Bot;
 use App\Domain\Conversation\Conversation;
 use App\Domain\Conversation\ConversationRepository;
-use App\Domain\Bot\Service\ChatPromptService;
-use App\Domain\Bot\Service\WebSearchInterface;
-use App\Domain\Bot\Service\GptInterface;
+use App\Domain\Bot\Service\ChatService;
 use App\Application\BotResponse;
 use App\Domain\Bot\ValueObject\Message;
-use App\Domain\Bot\Messages;
 
 class DefaultChatHandler implements CommandHandlerInterface
 {
-    private GptInterface $gpt;
+    private ChatService $chatService;
     private ConversationRepository $conversationRepository;
-    private ChatPromptService $chatPromptService;
-    private ?WebSearchInterface $webSearchTool;
-
-    const RECENT_CONVERSATIONS_COUNT_FOR_GPT = 10;
 
     public function __construct(
-        GptInterface $gpt,
-        ConversationRepository $conversationRepository,
-        ChatPromptService $chatPromptService,
-        ?WebSearchInterface $webSearchTool = null
+        ChatService $chatService,
+        ConversationRepository $conversationRepository
     ) {
-        $this->gpt = $gpt;
+        $this->chatService = $chatService;
         $this->conversationRepository = $conversationRepository;
-        $this->chatPromptService = $chatPromptService;
-        $this->webSearchTool = $webSearchTool;
     }
 
     public function canHandle(Command $command): bool
@@ -43,7 +32,7 @@ class DefaultChatHandler implements CommandHandlerInterface
 
     public function handle(Message $message, Bot $bot, Command $command): BotResponse
     {
-        $answer = $this->getAnswer($bot, $message);
+        $answer = $this->chatService->generateAnswer($bot, $message);
 
         // Avoid storing system-triggered messages (e.g., timer executions) in conversation history.
         if (!$message->isSystem()) {
@@ -51,44 +40,6 @@ class DefaultChatHandler implements CommandHandlerInterface
         }
 
         return new BotResponse($answer);
-    }
-
-    private function getAnswer(Bot $bot, Message $message): string
-    {
-        $recentConversations = $this->conversationRepository->findByBotId(
-            $bot->getId(),
-            self::RECENT_CONVERSATIONS_COUNT_FOR_GPT
-        );
-
-        $webSearchResults = null;
-        if ($this->shouldPerformWebSearch($message->getContent())) {
-            if ($this->webSearchTool instanceof WebSearchInterface) {
-                $webSearchResults = $this->webSearchTool->search($message->getContent(), 5);
-            } else {
-                $webSearchResults = "Error: Web search tool is not configured properly or failed to initialize.";
-            }
-        }
-
-        $configRequests = $bot->getConfigRequests(usePersonal: true, useDefault: true);
-
-        return $this->gpt->getAnswer(
-            context: $this->chatPromptService->generateContext(
-                $bot,
-                $recentConversations,
-                $configRequests,
-                $webSearchResults
-            ),
-            message: $message->getContent(),
-        );
-    }
-
-    private function shouldPerformWebSearch(string $messageContent): bool
-    {
-        $response = trim($this->gpt->getAnswer(
-            context: Messages::PROMPT_JUDGE_WEB_SEARCH,
-            message: $messageContent,
-        ));
-        return $response === "はい";
     }
 
     private function storeConversations(Bot $bot, string $messageContent, string $answer): void
